+++ packages/web-shared/src/components/workflow-trace-view.tsx
@@ -17,6 +17,7 @@ import type { MouseEvent as ReactMouseEvent, ReactNode } from 'react';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { useToast } from '../lib/toast';
+import { buildTrace, type TraceWithMeta } from '../lib/trace-builder';
 import { ErrorBoundary } from './error-boundary';
 import {
   EntityDetailPanel,
@@ -36,12 +37,6 @@ import {
   getCustomSpanClassName,
   getCustomSpanEventClassName,
 } from './workflow-traces/trace-colors';
-import { buildTrace, type TraceWithMeta } from '../lib/trace-builder';
-import { NewTraceViewer } from './trace-viewer-new';
-import {
-  SidebarDataProvider,
-  type SidebarDataContextValue,
-} from './sidebar/sidebar-data-context';
 
 /**
  * While a run is live, continuously grow root.duration and rescale so the
@@ -861,39 +856,6 @@ export const WorkflowTraceViewer = ({
     [onSpanSelect]
   );
 
-  const sidebarData = useMemo<SidebarDataContextValue>(
-    () => ({
-      run,
-      events,
-      spanDetailData: spanDetailData ?? null,
-      spanDetailError,
-      spanDetailLoading,
-      onSpanSelect: handleSpanSelect,
-      onStreamClick,
-      onWakeUpSleep,
-      onLoadEventData,
-      onResolveHook,
-      encryptionKey,
-      onDecrypt,
-      isDecrypting,
-    }),
-    [
-      run,
-      events,
-      spanDetailData,
-      spanDetailError,
-      spanDetailLoading,
-      handleSpanSelect,
-      onStreamClick,
-      onWakeUpSleep,
-      onLoadEventData,
-      onResolveHook,
-      encryptionKey,
-      onDecrypt,
-      isDecrypting,
-    ]
-  );
-
   const handleSelectionChange = useCallback(
     (info: SelectedSpanInfo | null) => {
       if (info) {
@@ -1004,9 +966,6 @@ export const WorkflowTraceViewer = ({
 
   return (
     <div className="relative w-full h-full flex flex-col">
-      <SidebarDataProvider value={sidebarData}>
-        <NewTraceViewer run={run} events={events} />
-      </SidebarDataProvider>
       {/* Timeline (takes remaining space) */}
       <div className="flex-1 min-w-0 relative">
         <TraceViewerContextProvider
