+++ packages/web/app/components/run-detail-view.tsx
@@ -6,9 +6,11 @@ import {
   EventListView,
   hydrateResourceIO,
   hydrateResourceIOWithKey,
+  NewTraceViewer,
+  type SidebarDataContextValue,
+  SidebarDataProvider,
   StreamViewer,
   stepEventsToStepEntity,
-  WorkflowTraceViewer,
 } from '@workflow/web-shared';
 import type { Event, WorkflowRun } from '@workflow/world';
 import {
@@ -18,7 +20,6 @@ import {
   List,
   Loader2,
   Lock,
-  Unlock,
 } from 'lucide-react';
 import { useCallback, useMemo, useRef, useState } from 'react';
 import { Link, useNavigate, useSearchParams } from 'react-router';
@@ -42,20 +43,17 @@ import {
   BreadcrumbPage,
   BreadcrumbSeparator,
 } from '~/components/ui/breadcrumb';
-import { Button } from '~/components/ui/button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '~/components/ui/tabs';
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
 } from '~/components/ui/tooltip';
+import { useEventsListData } from '~/lib/client/hooks/use-events-list-data';
 import { mapRunToExecution } from '~/lib/flow-graph/graph-execution-mapper';
 import { useWorkflowGraphManifest } from '~/lib/flow-graph/use-workflow-graph';
 import { useStreamReader } from '~/lib/hooks/use-stream-reader';
-
 import { fetchEvent, getEncryptionKeyForRun } from '~/lib/rpc-client';
-
-import { useEventsListData } from '~/lib/client/hooks/use-events-list-data';
 import type { EnvMap } from '~/lib/types';
 import {
   cancelRun,
@@ -338,9 +336,6 @@ export function RunDetailView({
     loading,
     error,
     update,
-    loadMoreTraceData,
-    hasMoreTraceData,
-    isLoadingMoreTraceData,
   } = useWorkflowTraceViewerData(env, runId, { live: true });
 
   const run = runData ?? ({} as WorkflowRun);
@@ -417,6 +412,39 @@ export function RunDetailView({
   const handleSpanSelect = useCallback((info: SpanSelectionInfo) => {
     setSpanSelection(info);
   }, []);
+
+  const sidebarData = useMemo<SidebarDataContextValue>(
+    () => ({
+      run,
+      events: allEvents ?? [],
+      spanDetailData: spanDetailData ?? null,
+      spanDetailError,
+      spanDetailLoading,
+      onSpanSelect: handleSpanSelect,
+      onStreamClick: handleStreamClick,
+      onWakeUpSleep: handleWakeUpSleep,
+      onLoadEventData: handleLoadSidebarEventData,
+      onResolveHook: handleResolveHook,
+      encryptionKey: encryptionKey ?? undefined,
+      onDecrypt: handleDecrypt,
+      isDecrypting,
+    }),
+    [
+      run,
+      allEvents,
+      spanDetailData,
+      spanDetailError,
+      spanDetailLoading,
+      handleSpanSelect,
+      handleStreamClick,
+      handleWakeUpSleep,
+      handleLoadSidebarEventData,
+      handleResolveHook,
+      encryptionKey,
+      handleDecrypt,
+      isDecrypting,
+    ]
+  );
 
   // Fetch streams for this run
   const {
@@ -741,26 +769,9 @@ export function RunDetailView({
             <TabsContent value="trace" className="mt-0 flex-1 min-h-0">
               <ErrorBoundary title="Failed to load trace viewer">
                 <div className="h-full">
-                  <WorkflowTraceViewer
-                    error={error}
-                    events={allEvents}
-                    run={run}
-                    isLoading={loading}
-                    spanDetailData={spanDetailData}
-                    spanDetailLoading={spanDetailLoading}
-                    spanDetailError={spanDetailError}
-                    onSpanSelect={handleSpanSelect}
-                    onStreamClick={handleStreamClick}
-                    onWakeUpSleep={handleWakeUpSleep}
-                    onResolveHook={handleResolveHook}
-                    onLoadEventData={handleLoadSidebarEventData}
-                    onLoadMoreSpans={loadMoreTraceData}
-                    hasMoreSpans={hasMoreTraceData}
-                    isLoadingMoreSpans={isLoadingMoreTraceData}
-                    encryptionKey={encryptionKey ?? undefined}
-                    onDecrypt={handleDecrypt}
-                    isDecrypting={isDecrypting}
-                  />
+                  <SidebarDataProvider value={sidebarData}>
+                    <NewTraceViewer run={run} events={allEvents ?? []} />
+                  </SidebarDataProvider>
                 </div>
               </ErrorBoundary>
             </TabsContent>
